requirements.txt
# Database dependencies
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pandas==2.1.4

# API clients
google-ads==22.1.0
bingads==13.0.16

# Environment and configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Scheduling and threading
schedule==1.2.1

# Web server and dashboard
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.1

# Data processing
numpy==1.26.3

# Utilities
requests==2.31.0
